requirements.txt
torch
pandas 
scikit-learn
matplotlib
tqdm
ipykernel
biopython
tensorboard
